requirements.txt
customtkinter==5.2.1
pycryptodome==3.20.0
